requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
requests-oauthlib==2.0.0
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic==2.7.1
firebase-admin==6.5.0
cryptography==42.0.5
